Resolves GH-23: Switch to constant-time check in WebhookVerifier
Previously, WebhookVerifier delegated to String.equals(Object) to
compare an expected security key and a provided one. This left code
utilizing that implementation more vulnerable to timing attacks, as the
time required to make the comparison varied based on how close to
correct the provided key was.

This change switches the implementation to use
MessageDigest.isEqual(byte[], byte[]), which since Java SE 6u17 has used
a constant-time implementation to reduce the risk of such attacks