Resolves GH-34 Configure GitHub authentication to reduce timing errors
Previously, client would sometimes observe 401 responses when attempting
to utilize the access token JWT in requests to GitHub. This can be
caused if GitHub believes an expiration time of greater than 10 minutes
was requested.

To reduce the possibility of this error, this changeset makes two
adjustments. The first is to reduce the default caching and expiration
times of both the access and installation tokesn by one minute, to
decrease the possible timing issues when communicating wiht GitHub. The
second is to make these expiration times configurable, so that clients
with less reliable clocks can further reduce these times themselves if
necessary